requirements.txt
Flask
Werkzeug
Pillow
numpy
torch
torchvision
pycryptodome
opencv-python
tqdm
matplotlib
requests